requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11
django-allauth==0.38.0
django-celery==3.2.2
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-formtools==1.0
django-import-export==1.1.0
django-js-asset==1.1.0
django-reversion==3.0.2
DjangoUeditor==1.8.143
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.24.1
future==0.15.2
greenlet==0.4.15
httplib2==0.9.2
idna==2.7
itchat==1.3.10
jdcal==1.4
kombu==3.0.37
lxml==4.2.5
monotonic==1.5
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.12
Pillow==5.3.0
pycparser==2.19
PyDocX==0.9.10
PyMySQL==0.9.2
pypng==0.0.18
PyQRCode==1.2.1
python-docx==0.8.7
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
redis==2.10.6
requests==2.20.1
requests-oauthlib==1.0.0
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
vine==1.2.0
xlrd==1.1.0
xlwt==1.3.0
